requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
joblib>=1.3.0

# 데이터 수집 (02_instagram_data_analysis)
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
lxml>=4.9.0
